+++ tests/unit/experimental/components/test_entity_relation_extractor.py
@@ -20,12 +20,13 @@ from unittest.mock import MagicMock
 import pytest
 from neo4j_genai.exceptions import LLMGenerationError
 from neo4j_genai.experimental.components.entity_relation_extractor import (
-    EntityRelationExtractor,
+    LexicalGraphBuilder,
     LLMEntityRelationExtractor,
     OnError,
     balance_curly_braces,
     fix_invalid_json,
 )
+from neo4j_genai.experimental.components.pdf_loader import DocumentInfo
 from neo4j_genai.experimental.components.types import (
     Neo4jGraph,
     Neo4jNode,
@@ -36,42 +37,59 @@ from neo4j_genai.llm import LLMInterface, LLMResponse
 
 
 def test_create_chunk_node_no_metadata() -> None:
-    # instantiating an abstract class to test common methods
-    extractor = EntityRelationExtractor()  # type: ignore
-    node = extractor.create_chunk_node(
-        chunk=TextChunk(text="text chunk"), chunk_id="10"
+    builder = LexicalGraphBuilder()
+    node = builder.create_chunk_node(
+        chunk=TextChunk(text="text chunk", index=0), chunk_id="10"
     )
     assert isinstance(node, Neo4jNode)
     assert node.id == "10"
-    assert node.properties == {"text": "text chunk"}
+    assert node.properties == {"index": 0, "text": "text chunk"}
     assert node.embedding_properties == {}
 
 
 def test_create_chunk_node_metadata_no_embedding() -> None:
-    # instantiating an abstract class to test common methods
-    extractor = EntityRelationExtractor()  # type: ignore
-    node = extractor.create_chunk_node(
-        chunk=TextChunk(text="text chunk", metadata={"status": "ok"}), chunk_id="10"
+    builder = LexicalGraphBuilder()
+    node = builder.create_chunk_node(
+        chunk=TextChunk(text="text chunk", index=0, metadata={"status": "ok"}),
+        chunk_id="10",
     )
     assert isinstance(node, Neo4jNode)
     assert node.id == "10"
-    assert node.properties == {"text": "text chunk", "status": "ok"}
+    assert node.properties == {"index": 0, "text": "text chunk", "status": "ok"}
     assert node.embedding_properties == {}
 
 
 def test_create_chunk_node_metadata_embedding() -> None:
-    # instantiating an abstract class to test common methods
-    extractor = EntityRelationExtractor()  # type: ignore
-    node = extractor.create_chunk_node(
+    builder = LexicalGraphBuilder()
+    node = builder.create_chunk_node(
         chunk=TextChunk(
-            text="text chunk", metadata={"status": "ok", "embedding": [1, 2, 3]}
+            text="text chunk",
+            index=0,
+            metadata={"status": "ok", "embedding": [1, 2, 3]},
         ),
         chunk_id="10",
     )
     assert isinstance(node, Neo4jNode)
     assert node.id == "10"
-    assert node.properties == {"text": "text chunk", "status": "ok"}
+    assert node.properties == {"index": 0, "text": "text chunk", "status": "ok"}
     assert node.embedding_properties == {"embedding": [1, 2, 3]}
+
+
+@pytest.mark.asyncio
+async def test_extractor_happy_path_no_entities_no_document() -> None:
+    llm = MagicMock(spec=LLMInterface)
+    llm.ainvoke.return_value = LLMResponse(content='{"nodes": [], "relationships": []}')
+
+    extractor = LLMEntityRelationExtractor(
+        llm=llm,
+    )
+    chunks = TextChunks(chunks=[TextChunk(text="some text", index=0)])
+    result = await extractor.run(chunks=chunks)
+    assert isinstance(result, Neo4jGraph)
+    # only one Chunk node (no document info provided)
+    assert len(result.nodes) == 1
+    assert result.nodes[0].label == "Chunk"
+    assert result.relationships == []
 
 
 @pytest.mark.asyncio
@@ -82,13 +100,15 @@ async def test_extractor_happy_path_no_entities() -> None:
     extractor = LLMEntityRelationExtractor(
         llm=llm,
     )
-    chunks = TextChunks(chunks=[TextChunk(text="some text")])
-    result = await extractor.run(chunks=chunks)
+    chunks = TextChunks(chunks=[TextChunk(text="some text", index=0)])
+    document_info = DocumentInfo(path="path")
+    result = await extractor.run(chunks=chunks, document_info=document_info)
     assert isinstance(result, Neo4jGraph)
-    # only one Chunk node
-    assert len(result.nodes) == 1
-    assert result.nodes[0].label == "Chunk"
-    assert result.relationships == []
+    # one Chunk node and one Document node
+    assert len(result.nodes) == 2
+    assert set(n.label for n in result.nodes) == {"Chunk", "Document"}
+    assert len(result.relationships) == 1
+    assert result.relationships[0].type == "FROM_DOCUMENT"
 
 
 @pytest.mark.asyncio
@@ -100,8 +120,9 @@ async def test_extractor_happy_path_no_entities_no_lexical_graph() -> None:
         llm=llm,
         create_lexical_graph=False,
     )
-    chunks = TextChunks(chunks=[TextChunk(text="some text")])
-    result = await extractor.run(chunks=chunks)
+    chunks = TextChunks(chunks=[TextChunk(text="some text", index=0)])
+    document_info = DocumentInfo(path="path")
+    result = await extractor.run(chunks=chunks, document_info=document_info)
     assert result.nodes == []
     assert result.relationships == []
 
@@ -116,18 +137,24 @@ async def test_extractor_happy_path_non_empty_result() -> None:
     extractor = LLMEntityRelationExtractor(
         llm=llm,
     )
-    chunks = TextChunks(chunks=[TextChunk(text="some text")])
-    result = await extractor.run(chunks=chunks)
+    chunks = TextChunks(chunks=[TextChunk(text="some text", index=0)])
+    document_info = DocumentInfo(path="path")
+    result = await extractor.run(chunks=chunks, document_info=document_info)
     assert isinstance(result, Neo4jGraph)
-    assert len(result.nodes) == 2
-    entity = result.nodes[0]
+    assert len(result.nodes) == 3
+    doc = result.nodes[0]
+    assert doc.label == "Document"
+    entity = result.nodes[1]
     assert entity.id.endswith("0:0")
     assert entity.label == "Person"
     assert entity.properties == {"chunk_index": 0}
-    chunk_entity = result.nodes[1]
+    chunk_entity = result.nodes[2]
     assert chunk_entity.label == "Chunk"
-    assert len(result.relationships) == 1
-    assert result.relationships[0].type == "FROM_CHUNK"
+    assert len(result.relationships) == 2
+    assert result.relationships[0].type == "FROM_DOCUMENT"
+    assert result.relationships[0].start_node_id.endswith(":0")
+    assert result.relationships[0].end_node_id == "path"
+    assert result.relationships[1].type == "FROM_CHUNK"
 
 
 @pytest.mark.asyncio
@@ -139,7 +166,7 @@ async def test_extractor_missing_entity_id() -> None:
     extractor = LLMEntityRelationExtractor(
         llm=llm,
     )
-    chunks = TextChunks(chunks=[TextChunk(text="some text")])
+    chunks = TextChunks(chunks=[TextChunk(text="some text", index=0)])
     with pytest.raises(LLMGenerationError):
         await extractor.run(chunks=chunks)
 
@@ -152,7 +179,7 @@ async def test_extractor_llm_ainvoke_failed() -> None:
     extractor = LLMEntityRelationExtractor(
         llm=llm,
     )
-    chunks = TextChunks(chunks=[TextChunk(text="some text")])
+    chunks = TextChunks(chunks=[TextChunk(text="some text", index=0)])
     with pytest.raises(LLMGenerationError):
         await extractor.run(chunks=chunks)
 
@@ -167,7 +194,7 @@ async def test_extractor_llm_badly_formatted_json() -> None:
     extractor = LLMEntityRelationExtractor(
         llm=llm,
     )
-    chunks = TextChunks(chunks=[TextChunk(text="some text")])
+    chunks = TextChunks(chunks=[TextChunk(text="some text", index=0)])
     with pytest.raises(LLMGenerationError):
         await extractor.run(chunks=chunks)
 
@@ -185,7 +212,7 @@ async def test_extractor_llm_invalid_json() -> None:
     extractor = LLMEntityRelationExtractor(
         llm=llm,
     )
-    chunks = TextChunks(chunks=[TextChunk(text="some text")])
+    chunks = TextChunks(chunks=[TextChunk(text="some text", index=0)])
     with pytest.raises(LLMGenerationError):
         await extractor.run(chunks=chunks)
 
@@ -202,7 +229,7 @@ async def test_extractor_llm_badly_formatted_json_do_not_raise() -> None:
         on_error=OnError.IGNORE,
         create_lexical_graph=False,
     )
-    chunks = TextChunks(chunks=[TextChunk(text="some text")])
+    chunks = TextChunks(chunks=[TextChunk(text="some text", index=0)])
     res = await extractor.run(chunks=chunks)
     assert res.nodes == []
     assert res.relationships == []